Improved performance for VSEncodingSimple.cpp

+++ tool/vcompress.cpp
@@ -383,7 +383,8 @@ void do_decompress(const std::string& input,
 
   /* Show performance results */
   fprintf(stdout, "Performance Results:\n");
-  fprintf(stdout, "  Total Num Decoded: %llu\n", dnum);
+  fprintf(stdout, "  Total Num Decoded: %llu\n",
+          static_cast<unsigned long long>(dnum));
   fprintf(stdout, "  Elapsed: %.2lf\n", elapsed);
   fprintf(stdout, "  Performance: %.2lfmis\n",
           (dnum + 0.0) / (elapsed * 1000000));
